# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "This API provides endpoints for the various users of the Reward Loyalty script.",
        "contact": {
            "email": "support@nowsquare.com"
        },
        "version": "1.6.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://reward-loyalty-demo.nowsquare.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/member/register": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Register a new member",
                "description": "Allows a new member to create an account by providing necessary details.",
                "operationId": "registerMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "email@example.com"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 64,
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the member, if empty a password is generated",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 48,
                            "minLength": 6,
                            "example": "password123"
                        }
                    },
                    {
                        "name": "time_zone",
                        "in": "query",
                        "description": "The time zone of the member",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "America/New_York"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "The locale of the member (e.g., `en_US`)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 12,
                            "minLength": 5,
                            "example": "en_US"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "The currency of the member (e.g., `USD`)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 3,
                            "minLength": 3,
                            "example": "USD"
                        }
                    },
                    {
                        "name": "accepts_emails",
                        "in": "query",
                        "description": "Whether the member accepts marketing emails from a partner or not",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 0
                        }
                    },
                    {
                        "name": "send_mail",
                        "in": "query",
                        "description": "Whether to send an email with the password to the newly created member",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberRegistration"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{locale}/v1/member/login": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Log in a member",
                "description": "Authenticates a member using their email and password.",
                "operationId": "loginMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g. `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the member",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password of the member",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberLoginSuccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/member/logout": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Log out a member",
                "description": "Revokes all access tokens for the authenticated member, effectively logging them out.",
                "operationId": "logoutMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/member": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "Retrieve member data",
                "description": "Fetches the data of the authenticated member.",
                "operationId": "getMember",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "Standard error response structure",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    },
                    "code": {
                        "description": "HTTP error code",
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
            "UnauthenticatedResponse": {
                "description": "Response structure when user is unauthenticated",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "description": "Response structure for validation errors",
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validation error occurred."
                    },
                    "errors": {
                        "description": "Detailed validation errors",
                        "properties": {
                            "email": {
                                "description": "List of email validation errors",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "The email must be a valid email address."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "MemberLoginSuccess": {
                "required": [
                    "token"
                ],
                "properties": {
                    "token": {
                        "description": "Generated authentication token for the member",
                        "type": "string",
                        "example": "1|laravel_sanctum_j3ffUNZdoP0JZJn0y3GzgcAONlDzekQrUWI2sqk3c473f47b"
                    }
                },
                "type": "object"
            },
            "MemberRegistration": {
                "required": [],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "email@example.com"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "John Doe"
                    },
                    "password": {
                        "description": "Password for the member",
                        "type": "string",
                        "maxLength": 48,
                        "example": "mypassword",
                        "nullable": true
                    },
                    "time_zone": {
                        "description": "Time zone of the member",
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "locale": {
                        "description": "Locale of the member",
                        "type": "string",
                        "maxLength": 12,
                        "minLength": 5,
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "description": "Preferred currency of the member",
                        "type": "string",
                        "maxLength": 3,
                        "minLength": 3,
                        "example": "USD",
                        "nullable": true
                    },
                    "accepts_emails": {
                        "description": "Indicates if the member accepts emails",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 1,
                        "nullable": true
                    },
                    "send_mail": {
                        "description": "If set to 1, sends an email with the password to the newly registered member",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 0,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Member": {
                "required": [],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 81559379910656
                    },
                    "unique_identifier": {
                        "description": "Unique identifier for the member",
                        "type": "string",
                        "example": "700-857-223-945"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 64,
                        "example": "Member Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "member@example.com"
                    },
                    "email_verified_at": {
                        "description": "Timestamp when the email was verified",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-20T09:12:01.000000Z"
                    },
                    "locale": {
                        "description": "Locale of the member",
                        "type": "string",
                        "maxLength": 12,
                        "minLength": 5,
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "description": "Preferred currency of the member",
                        "type": "string",
                        "maxLength": 3,
                        "minLength": 3,
                        "example": "USD",
                        "nullable": true
                    },
                    "time_zone": {
                        "description": "Time zone of the member",
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "accepts_emails": {
                        "description": "Indicates if the member accepts emails",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 0,
                        "nullable": true
                    },
                    "number_of_times_logged_in": {
                        "description": "Number of times the member has logged in",
                        "type": "integer",
                        "example": 6
                    },
                    "last_login_at": {
                        "description": "Last login timestamp of the member",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-13T13:20:57.000000Z"
                    },
                    "created_at": {
                        "description": "Creation timestamp of the member record",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-20T09:12:01.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the member record",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-13T13:20:57.000000Z"
                    },
                    "avatar": {
                        "description": "Avatar URL of the member",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Admin authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token, formatted as `1|laravel_sanctum_<token_string>`.",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}